# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Trading Journal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /instruments:
    get:
      operationId: listInstruments
      tags: [instruments]
      summary: List instruments
      parameters:
        - name: assetClass
          in: query
          schema:
            type: string
        - name: isActive
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of instruments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Instrument"
    post:
      operationId: createInstrument
      tags: [instruments]
      summary: Create an instrument
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InstrumentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Instrument"
  /instruments/{id}:
    get:
      operationId: getInstrument
      tags: [instruments]
      summary: Get instrument by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Instrument
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Instrument"
        "404":
          description: Not found
    patch:
      operationId: updateInstrument
      tags: [instruments]
      summary: Update instrument
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InstrumentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Instrument"
        "404":
          description: Not found
    delete:
      operationId: deleteInstrument
      tags: [instruments]
      summary: Delete instrument
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # ─── STRATEGIES ─────────────────────────────────────────────────────────────
  /strategies:
    get:
      operationId: listStrategies
      tags: [strategies]
      summary: List strategies
      responses:
        "200":
          description: List of strategies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Strategy"
    post:
      operationId: createStrategy
      tags: [strategies]
      summary: Create a strategy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StrategyInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Strategy"
  /strategies/{id}:
    get:
      operationId: getStrategy
      tags: [strategies]
      summary: Get strategy by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Strategy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Strategy"
        "404":
          description: Not found
    patch:
      operationId: updateStrategy
      tags: [strategies]
      summary: Update strategy
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StrategyUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Strategy"
        "404":
          description: Not found
    delete:
      operationId: deleteStrategy
      tags: [strategies]
      summary: Delete strategy
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # ─── SETUPS ─────────────────────────────────────────────────────────────────
  /setups:
    get:
      operationId: listSetups
      tags: [setups]
      summary: List setups
      parameters:
        - name: strategyId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of setups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Setup"
    post:
      operationId: createSetup
      tags: [setups]
      summary: Create a setup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetupInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Setup"
  /setups/{id}:
    patch:
      operationId: updateSetup
      tags: [setups]
      summary: Update setup
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetupUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Setup"
        "404":
          description: Not found
    delete:
      operationId: deleteSetup
      tags: [setups]
      summary: Delete setup
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # ─── CONFLUENCE CATEGORIES ───────────────────────────────────────────────────
...
  /tags:
    get:
      operationId: listTags
      tags: [tags]
      summary: List all tags
      responses:
        "200":
          description: List of tags
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tag"
    post:
      operationId: createTag
      tags: [tags]
      summary: Create a tag
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TagInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tag"
  /tags/{id}:
    delete:
      operationId: deleteTag
      tags: [tags]
      summary: Delete a tag
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # ─── TRADES ─────────────────────────────────────────────────────────────────
...
  /trades/{tradeId}/review:
    get:
      operationId: getTradeReview
      tags: [reviews]
      summary: Get the review for a trade
      parameters:
        - name: tradeId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Review (may be empty if not yet written)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TradeReview"
        "404":
          description: Trade not found
    put:
      operationId: upsertTradeReview
      tags: [reviews]
      summary: Create or update the review for a trade
      parameters:
        - name: tradeId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TradeReviewInput"
      responses:
        "200":
          description: Review upserted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TradeReview"
        "404":
          description: Trade not found

  # ─── DASHBOARD ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Instrument:
      type: object
      required: [id, symbol, name, assetClass, settlementCurrency, quantityUnit, isActive, createdAt, updatedAt]
      properties:
        id: { type: integer }
        symbol: { type: string }
        name: { type: string }
        assetClass: { type: string }
        settlementCurrency: { type: string }
        quantityUnit: { type: string }
        contractMultiplier: { type: ["number", "null"] }
        isActive: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    InstrumentInput:
      type: object
      required: [symbol, name, assetClass, settlementCurrency, quantityUnit]
      properties:
        symbol: { type: string, minLength: 1 }
        name: { type: string, minLength: 1 }
        assetClass: { type: string, minLength: 1 }
        settlementCurrency: { type: string, minLength: 1 }
        quantityUnit: { type: string, minLength: 1 }
        contractMultiplier: { type: number }
        isActive: { type: boolean, default: true }
    InstrumentUpdate:
      type: object
      properties:
        symbol: { type: string }
        name: { type: string }
        assetClass: { type: string }
        settlementCurrency: { type: string }
        quantityUnit: { type: string }
        contractMultiplier: { type: ["number", "null"] }
        isActive: { type: boolean }

    # ── STRATEGY ─────────────────────────────────────────────────────────────
    Strategy:
      type: object
      required: [id, name, isActive, createdAt, updatedAt]
      properties:
        id: { type: integer }
        name: { type: string }
        description: { type: ["string", "null"] }
        isActive: { type: boolean }
        setups:
          type: array
          items:
            $ref: "#/components/schemas/Setup"
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    StrategyInput:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1 }
        description: { type: string }
        isActive: { type: boolean, default: true }
    StrategyUpdate:
      type: object
      properties:
        name: { type: string }
        description: { type: ["string", "null"] }
        isActive: { type: boolean }

    # ── SETUP ────────────────────────────────────────────────────────────────
    Setup:
      type: object
      required: [id, name, isActive, createdAt, updatedAt]
      properties:
        id: { type: integer }
        strategyId: { type: ["integer", "null"] }
        strategyName: { type: ["string", "null"] }
        name: { type: string }
        description: { type: ["string", "null"] }
        isActive: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    SetupInput:
      type: object
      required: [name]
      properties:
        strategyId: { type: integer }
        name: { type: string, minLength: 1 }
        description: { type: string }
        isActive: { type: boolean, default: true }
    SetupUpdate:
      type: object
      properties:
        strategyId: { type: ["integer", "null"] }
        name: { type: string }
        description: { type: ["string", "null"] }
        isActive: { type: boolean }

    # ── CONFLUENCE CATEGORY ──────────────────────────────────────────────────
...
    Tag:
      type: object
      required: [id, name, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        createdAt: { type: string, format: date-time }
    TagInput:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1 }

    # ── EXECUTION ────────────────────────────────────────────────────────────
...
    TradeReview:
      type: object
      required: [id, tradeId, createdAt, updatedAt]
      properties:
        id: { type: integer }
        tradeId: { type: integer }
        reviewNotes: { type: ["string", "null"] }
        whatWentWell: { type: ["string", "null"] }
        whatWentWrong: { type: ["string", "null"] }
        ruleAdherence: { type: ["integer", "null"] }
        tradeQuality: { type: ["integer", "null"] }
        lessonsLearned: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    TradeReviewInput:
      type: object
      properties:
        reviewNotes: { type: ["string", "null"] }
        whatWentWell: { type: ["string", "null"] }
        whatWentWrong: { type: ["string", "null"] }
        ruleAdherence: { type: ["integer", "null"] }
        tradeQuality: { type: ["integer", "null"] }
        lessonsLearned: { type: ["string", "null"] }

    # ── TRADE CALCULATED VALUES ───────────────────────────────────────────────
...