# lib/api-zod/src/generated/types/strategyInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Trading Journal API
 * OpenAPI spec version: 0.1.0
 */

export interface StrategyInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  isActive?: boolean;
}